requirements.txt
###### Requirements without Version Specifiers ######`
scipy.stats
matplotlib
pandas

